provide an empty import method in the generated module
During development, the inline module is meant to be passed C code to
compile. For a release, it is replaced with a module that already has
the code compiled, so its import is meant to do nothing.

Future versions of perl are planned to throw errors if an undefined
import method is given arguments, since this generally indicates a
mistake.

Provide an empty import method to explicitly ignore the arguments.